[Refactor] #179 - Intent State Store에 통합 | 불변 객체 State와 분리
